+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Add tooltip to search results' title
+- Automatically resume video from where the user left off
 
 ### Fixed
 - Handle posts without thumbnail properly